app/retrieval.py
import faiss
import numpy as np
from sentence_transformers import SentenceTransformer

# Placeholder for FAISS vector store setup and query retrieval
def create_vector_store(documents):
    model = SentenceTransformer('all-MiniLM-L6-v2')
    embeddings = [model.encode(doc) for doc in documents]
    vector_store = faiss.IndexFlatL2(embeddings[0].shape[0])
    for emb in embeddings:
        vector_store.add(np.array([emb]))
    return vector_store

def retrieve_top_k(query, vector_store, documents, k=3):
    model = SentenceTransformer('all-MiniLM-L6-v2')
    query_emb = model.encode([query])
    distances, indices = vector_store.search(np.array(query_emb), k)
    return [documents[idx] for idx in indices[0]]

# Sample usage
documents = ["What is the warranty?", "How to return the product?", "Company background."]
vector_store = create_vector_store(documents)
top_k_docs = retrieve_top_k("warranty", vector_store, documents)
print(top_k_docs)